initial

+++ app/Http/Controllers/LegalEntityController.php
@@ -9,7 +9,11 @@ use Alert;
 use DB;
 use App\DataTables\LegalEntityDataTable;
 use App\Models\Company;
+use App\Models\Licensing;
 use Yajra\DataTables\Facades\DataTables;
+use Illuminate\Contracts\Encryption\DecryptException;
+use Illuminate\Support\Facades\Crypt;
+use Auth;
 
 class LegalEntityController extends Controller
 {
@@ -40,6 +44,7 @@ class LegalEntityController extends Controller
         $request->validate([
             'name'          => 'required|unique:legal_entities|min:2|max:255',
             'address'       => 'required|min:2|max:255',
+            'division'      => 'required',
             'description'   => 'max:500',
         ]);
 
@@ -56,7 +61,7 @@ class LegalEntityController extends Controller
             DB::commit();
             Alert::success('SUCCEED','Successfully save data to system');
 
-            return back();
+            return redirect()->route('index_legal');
         } catch (\Throwable $th) {
             //throw $th;
             DB::rollback();
@@ -69,9 +74,11 @@ class LegalEntityController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(LegalEntity $legalEntity)
+    public function show($id)
     {
-        //
+        $legal = LegalEntity::find(Crypt::decryptString($id));
+
+        return view('pages.legal.show',compact('legal'));
     }
 
     /**
@@ -79,7 +86,7 @@ class LegalEntityController extends Controller
      */
     public function edit(LegalEntity $legalEntity, $id)
     {
-        $legal = LegalEntity::find($id);
+        $legal = LegalEntity::find(Crypt::decryptString($id));
         $devision = Division::all();
 
         return view('pages.legal.edit',compact('legal','devision'));
@@ -127,7 +134,16 @@ class LegalEntityController extends Controller
         DB::beginTransaction();
 
         try {
-            LegalEntity::find($id)->delete();
+            $checkLegal = Licensing::where('legal_entity_id',$id)->first();
+
+            if($checkLegal) {
+                DB::rollback();
+                Alert::info('CANT','Deletion not allowed, has been used');
+
+                return back();
+            }else{
+                LegalEntity::find($id)->delete();
+            }
             
             DB::commit();
             Alert::success('SUCCEED','Data deletion has been successful');
@@ -136,8 +152,10 @@ class LegalEntityController extends Controller
         } catch (\Throwable $th) {
              DB::rollback();
 
-            Alert::error('FAIL','Failed to delete data, please check again');
-            return back();
+            // Alert::error('FAIL','Failed to delete data, please check again');
+            // return back();
+
+            return $th->getMessage();
         }
     }
 
@@ -149,7 +167,7 @@ class LegalEntityController extends Controller
         $model = LegalEntity::query();
             return DataTables::eloquent($model)
             ->addColumn('action',function($model){
-            $btn = '<div class="d-flex justify-content-center"><a href="javascript:void(0)" class="btn btn-primary btn-sm m-1"><i class="ti ti-eye"></i></a><button type="button" class="btn btn-danger btn-sm m-1" data-bs-toggle="modal" data-bs-target="#modalDelete'.$model->id.'"><i class="ti ti-trash"></i></button></div>
+            $btn = '<div class="d-flex justify-content-center"><a href="/legal/show/'.Crypt::encryptString($model->id).'" class="btn btn-primary btn-sm m-1"><i class="ti ti-eye"></i></a><a href="/legal/edit/'.Crypt::encryptString($model->id).'" class="btn btn-success btn-sm m-1" style="color: white"><i class="ti ti-pencil"></i></a><button type="button" class="btn btn-danger btn-sm m-1" data-bs-toggle="modal" data-bs-target="#modalDelete'.$model->id.'"><i class="ti ti-trash"></i></button></div>
                     <!-- Modal -->
                         <div class="modal fade" id="modalDelete'.$model->id.'" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
                         <div class="modal-dialog">
@@ -161,15 +179,17 @@ class LegalEntityController extends Controller
                                     <div class="col-4">
                                         <img src="https://cdn3d.iconscout.com/3d/premium/thumb/alert-notification-8400231-6672344.png" style="width: 100%;"/>
                                     </div>
-                                    <div class="col-8 text-center align-self-center text-uppercase">
-                                        <span>Konfirmasi Penghapusan Data Dari Layanan Legal</span><br>
-                                        <span class="fw-bold">'.$model->name.'</span>
+                                    <div class="col-8 text-sm">
+                                        <div class="from group mb-3">
+                                        <h5 class="mt-2">'.$model->name.'</h5>
+                                        </div>
+                                        <span style="font-size: 10px;" class="text-lowercase">Proses penghapusan data akan dilakukan secara permanen dari layanan legal, yakin '.Auth::user()->name.' ingin menghapus ?</span>
                                     </div>
                                 </div>
                             </div>
                             <div class="modal-footer">
-                                <button type="button" class="btn btn-secondary text-sm" data-bs-dismiss="modal">Batal</button>
-                                <a href="/legal/delete/'.$model->id.'" class="btn btn-danger text-sm">Hapus</a>
+                                <button type="button" class="btn btn-secondary text-sm" data-bs-dismiss="modal" style="border-radius: 50%;"><i class="bi bi-x fa-lg"></i></button>
+                                <a href="/legal/delete/'.$model->id.'" class="btn btn-primary text-sm text-capitalize" style="border-radius: 50%;"><i class="bi bi-check-lg fa-lg"></i></a>
                             </div>
                             </div>
                         </div>
@@ -177,7 +197,7 @@ class LegalEntityController extends Controller
                         ';
             return $btn;
         })
-        ->addColumn('divisions', '{{$model->legals->name}}')
+        ->addColumn('divisions', '{{$model->divisions->name}}')
         ->rawColumns(['action'])
         ->toJson();
     }
